[NO STORY] Separate generate and encrypt entity secret into two samples
