+++ README.md
@@ -6,15 +6,36 @@ To generate an entity secret and encrypt with the entity public key, and registe
 
 1. Choose a programming language: Select the programming language you are using for your application. We provide sample code snippets for Python and Golang.
 
-2. Follow the sample code to generate an entity secret.
+2. Use the sample code (`generate_hex_encoded_entity_secret`) to generate a hex-encoded entity secret. You can also generate a 32 byte data and hex-encode it by yourselves.
+
+Python
+```bash
+python python/generate_hex_encoded_entity_secret.py
+```
+
+Golang
+```bash
+go run golang/generate_hex_encoded_entity_secret.go
+```
+
 
 3. Acquire the entity public key: Use the provided API endpoint `GET /config/entity/publicKey` to obtain the entity public key securely. This public key is required for the encryption process.
 
-4. Input the acquired entity public key in the sample code, and run sample code to encrypt the entity secret and encode in base64, you will get **entity secret ciphertext** accordingly.
+4. Replace the entity public key and hex-encoded entity secret in the sample code (`generate_entity_secret_ciphertext`), the sample code will encrypt and encode the entity secret in base64, and you will get the **entity secret ciphertext** accordingly.
 
-5. Register the **entity secret ciphertext** in the Configurator Page in the developer dashboard and click Register. The entity secret only needs to be generated, encrypted, encoded and registered once, unless you need to rotate the entity secret.
+Python
+```bash
+python python/generate_entity_secret_ciphertext.py
+```
 
-6. Now you can append an **entity secret ciphertext** in the API requests. Note that the encryption and encoding of entity secret should be executed every time you append in an API request to prevent replay attack. There is no need to register an updated entity secret ciphertext unless you need to rotate the entity secret. Here’s the sample API request for reference:
+Golang
+```bash
+go run golang/generate_entity_secret_ciphertext.go
+```
+
+5. Register the **entity secret ciphertext** in the Configurator Page in the [developer dashboard](https://console.circle.com/wallets/dev/configurator) and click Register. The entity secret ciphertext only needs to be registered once, unless you need to rotate the entity secret.
+
+6. Now you can append an **entity secret ciphertext** in the API request body for developer-controlled wallets. Note that the encryption and encoding of entity secret needs to be executed every time you append in an API request to prevent replay attack. There is no need to register an updated entity secret ciphertext; simply use the entity secret ciphertext as a variable in your API request and obtain the latest ciphertext generated by rerunning the sample code (`generate_entity_secret_ciphertext`). Here’s the sample API request for reference: 
 
 ```bash
 curl --location  --request POST 'https://api.circle.com/v1/w3s/developer/walletSets' \
@@ -29,6 +50,6 @@ curl --location  --request POST 'https://api.circle.com/v1/w3s/developer/walletS
 
 **Note**: Make sure to install related libraries for encryption before using the sample code. For Python sample code please first `pip install pycryptodome`.
 
-**Note**: By running the sample code you will get a hex-encoded entity secret, please store the information carefully by yourself, as it is required for critical API requests and Circle does not store the information.
+**Note**: Please store the hex-encoded entity secret carefully by yourself, as it is required for critical API requests and Circle does not store the information.
 
 **Note**: Please refrain from directly embedding the hex-encoded entity secret within the code.
