+++ golang/generate_hex_encoded_entity_secret.go
@@ -1,0 +1,40 @@
+// Copyright (c) 2023, Circle Technologies, LLC. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package main is the entrypoint for the sample code.
+package main
+
+import (
+	"crypto/rand"
+	"fmt"
+	"io"
+)
+
+// GetEntropyCSPRNG creates a random entropy of 32 byte length.
+// This function generates a 32 byte random entity secret.
+func GetEntropyCSPRNG() []byte {
+	mainBuff := make([]byte, 32)
+	_, err := io.ReadFull(rand.Reader, mainBuff)
+	if err != nil {
+		panic("reading from crypto/rand failed: " + err.Error())
+	}
+	return mainBuff
+}
+
+// The following sample codes generate a distinct hex encoded entity secret with each execution
+// The generation of entity secret only need to be executed once unless you need to rotate entity secret.
+func main() {
+	entitySecret := GetEntropyCSPRNG()
+	fmt.Printf("Hex encoded entity secret: %x\n", entitySecret)
+}
